Policy-IAM
This IAM policy JSON file defines the permissions for a Lambda function to interact with an Amazon DynamoDB table. The policy grants essential actions on the specified DynamoDB resource (`CalculatorDatabase`) located in the `ap-southeast-2` region. 

Key Features:
Actions Allowed: 
  - `PutItem`: Insert items into the DynamoDB table.
  - `DeleteItem`: Remove items from the table.
  - `GetItem`: Retrieve specific items by primary key.
  - `Scan`: Read all items in the table for broad queries.
  - `Query`: Retrieve items based on specified key conditions.
  - `UpdateItem`: Modify existing items in the table.

Resource: The permissions are explicitly tied to the `CalculatorDatabase` table using its ARN (Amazon Resource Name) for secure and precise access control.

Usage: This policy ensures that the Lambda function has the required permissions to perform CRUD operations, making it ideal for serverless applications that require interaction with DynamoDB.

This policy can be attached to an IAM role to enable a Lambda function or other AWS services to securely access and manage data within the `CalculatorDatabase`.